Finalizar Compra

+++ src/Vendas/VendasService.java
@@ -14,7 +14,7 @@ public class VendasService {
     }
 
     // Registrar venda pelo nome do produto
-    public void registrarVenda(String nomeProduto, long quantidadeCliente) {
+    public void registrarVenda(String nomeProduto, float quantidadeCliente) {
         // Tentar vender o produto (atualiza estoque)
         boolean vendaSucesso = produtoService.vender(nomeProduto, quantidadeCliente);
